chore: rebuild the viewer bundle after rebasing onto main
