+++ resources/views/index.blade.php
@@ -8,7 +8,7 @@
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;700&family=JetBrains+Mono:wght@400;500;600&display=swap" rel="stylesheet" />
-    <script type="module" crossorigin src="/_laravel-brain/assets/index-BwWD5GvV.js"></script>
+    <script type="module" crossorigin src="/_laravel-brain/assets/index-CAYzS4wX.js"></script>
     <link rel="modulepreload" crossorigin href="/_laravel-brain/assets/rolldown-runtime-BHe-jwch.js">
     <link rel="modulepreload" crossorigin href="/_laravel-brain/assets/vendor-CsjAK7B8.js">
     <link rel="modulepreload" crossorigin href="/_laravel-brain/assets/vendor-react-CNktMmef.js">
